requirements.txt
torch
Pillow
numpy
torchvision
tqdm